@@ -3,52 +3,65 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace IdentityServer
 {
     public class Startup
     {
-        public Startup(IConfiguration configuration)
+        public Startup(IWebHostEnvironment environment, IConfiguration configuration)
         {
+            this.Environment = environment;
             this.Configuration = configuration;
         }
 
+        private IWebHostEnvironment Environment { get; }
+
         private IConfiguration Configuration { get; }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void ConfigureServices(IServiceCollection services)
         {
-            if (env.IsDevelopment())
+            services.AddControllersWithViews();
+
+            services.ConfigureDatabase(this.Configuration.GetConnectionString("DefaultConnection"));
+
+            services.AddAuthentication()
+                .AddGoogle(options =>
+                {
+                    // register your IdentityServer with Google at https://console.developers.google.com
+                    // enable the Google+ API
+                    // set the redirect URI to http://localhost:5000/signin-google
+                    options.ClientId = "copy client ID from Google here";
+                    options.ClientSecret = "copy client secret from Google here";
+                });
+        }
+
+        public void Configure(IApplicationBuilder app)
+        {
+            if (this.Environment.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
                 app.UseDatabaseErrorPage();
             }
             else
             {
-                app.UseExceptionHandler("/Home/Error");
                 app.UseHsts();
             }
 
-            app.UseHttpsRedirection();
             app.UseStaticFiles();
 
+            app.UseRouting();
             app.UseIdentityServer();
-
-            app.UseMvc(routes =>
+            app.UseAuthorization();
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
-                    name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                endpoints.MapDefaultControllerRoute();
             });
 
             app.InitializeDatabase(
                 this.Configuration["AdminClient:ApiName"],
                 this.Configuration["AdminClient:ClientId"],
                 this.Configuration["AdminClient:ClientSecret"]);
-        }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.ConfigureIdentity(this.Configuration.GetConnectionString("DefaultConnection"));
         }
     }
 }
