@@ -1,22 +1,25 @@
 ﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
+using IdentityModel;
 using IdentityServer.Admin.Authorization;
 using IdentityServer.Admin.Models;
 using IdentityServer.Common.Constants;
-using IdentityServer.Data;
+using IdentityServer.Data.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using SystemClaim = System.Security.Claims.Claim;
 
 namespace IdentityServer.Admin.Controllers
 {
     [ApiController]
-    [Route("[controller]")]
+    [Route("users")]
     [Authorize(Policy = Policies.ManageUsers)]
     public sealed class UsersController : BaseUserController
     {
-        public UsersController(UserManager<User> userManager, IMapper mapper)
+        public UsersController(UserManager<ApplicationUser> userManager, IMapper mapper)
             : base(userManager)
         {
             this.Mapper = mapper;
@@ -24,7 +27,7 @@ namespace IdentityServer.Admin.Controllers
 
         private IMapper Mapper { get; }
 
-        [HttpPost(Name = "CreateUser")]
+        [HttpPost]
         [Consumes(ContentTypes.ApplicationJson)]
         [Produces(ContentTypes.ApplicationJson)]
         [ProducesResponseType(200)]
@@ -34,14 +37,14 @@ namespace IdentityServer.Admin.Controllers
             if (request == null)
                 throw new ArgumentNullException(nameof(request));
 
-            var user = this.Mapper.Map<User>(request);
+            var user = this.Mapper.Map<ApplicationUser>(request);
 
             var result = await this.UserManager.CreateAsync(user, request.Password);
             return this.ConvertIdentityResultToResponse(result);
         }
 
         // TODO: allow a user to access their own details
-        [HttpGet("{email}", Name = "GetUser")]
+        [HttpGet("{email}")]
         [Produces(ContentTypes.ApplicationJson)]
         [ProducesResponseType(200, Type = typeof(UserDetails))]
         [ProducesResponseType(404)]
@@ -54,11 +57,20 @@ namespace IdentityServer.Admin.Controllers
             if (user == null)
                 return this.NotFound();
 
-            return this.Ok(this.Mapper.Map<UserDetails>(user));
+            var userDetails = this.Mapper.Map<UserDetails>(user);
+
+            var claims = await this.UserManager.GetClaimsAsync(user);
+            if (claims != null)
+            {
+                userDetails.GivenName = claims.FirstOrDefault(i => i.Type == JwtClaimTypes.GivenName)?.Value;
+                userDetails.Surname = claims.FirstOrDefault(i => i.Type == JwtClaimTypes.FamilyName)?.Value;
+            }
+
+            return this.Ok(userDetails);
         }
 
         // TODO: allow a user to update their own details
-        [HttpPut("{email}", Name = "UpdateUser")]
+        [HttpPut("{email}")]
         [Consumes(ContentTypes.ApplicationJson)]
         [Produces(ContentTypes.ApplicationJson)]
         [ProducesResponseType(200)]
@@ -78,12 +90,22 @@ namespace IdentityServer.Admin.Controllers
             if (user == null)
                 return this.NotFound();
 
-            user.GivenName = request.GivenName;
-            user.Surname = request.Surname;
+            var result = await this.UserManager.UpdateAsync(user);
 
-            await this.UserManager.UpdateAsync(user);
+            if (!result.Succeeded)
+                return this.ConvertIdentityResultToResponse(result);
 
-            return this.Ok();
+            var claims = new SystemClaim[]
+            {
+                new SystemClaim(JwtClaimTypes.Name, $"{request.GivenName} {request.Surname}"),
+                new SystemClaim(JwtClaimTypes.GivenName, request.GivenName),
+                new SystemClaim(JwtClaimTypes.FamilyName, request.Surname),
+                new SystemClaim(JwtClaimTypes.Email, request.Email)
+            };
+
+            result = await this.UserManager.AddClaimsAsync(user, claims);
+
+            return this.ConvertIdentityResultToResponse(result);
         }
     }
 }
