@@ -1,3 +1,2 @@
-﻿@using IdentityServer
-@using IdentityServer.Models
+﻿@using IdentityServer4.Quickstart.UI
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
